Moving some parts to shared module
Some auth parts that should not be lazy loaded are moved to shared module
