+++ FrontEnd/src/app/company/company-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { RolesEnum } from '../core/auth/enums/roles.enum';
-import { AuthGuard } from '../core/auth/guards/auth-guard';
+import { AuthGuard } from '../shared/guards/auth-guard';
 import { BoardCompanyComponent } from './components/board-company/board-company.component';
 
 const routes: Routes = [
