+++ FrontEnd/src/app/core/auth/components/login/login.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
+import { TokenStorageService } from 'src/app/shared/services/token-storage.service';
 import { LogInForm } from '../../forms/login.form';
 import { AuthService } from '../../services/auth.service';
-import { TokenStorageService } from '../../services/token-storage.service';
 
 @Component({
   templateUrl: './login.component.html',
