+++ FrontEnd/src/app/core/core.module.ts
@@ -1,7 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
-import { NavbarComponent } from './navbar/navbar.component';
 import { LogInComponent } from './auth/components/login/login.component';
 import { LogOutComponent } from './auth/components/logout/logout.component';
 import { RegisterComponent } from './auth/components/register/register.component';
@@ -9,10 +7,10 @@ import { ProfileComponent } from './auth/components/profile/profile.component';
 import { ChangePasswordComponent } from './auth/components/change-password/change-password.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
-import { authInterceptorProviders } from './auth/interceptors/auth.interceptor';
 import { SharedModule } from '../shared/shared.module';
 import { BoardUsersComponent } from './auth/components/board-users/board-users.component';
 import { AuthRoutingModule } from './auth/auth-routing.module';
+import { ErrorsRoutingModule } from './auth/errors-routing.module';
 
 @NgModule({
   declarations: [
@@ -22,7 +20,6 @@ import { AuthRoutingModule } from './auth/auth-routing.module';
     ProfileComponent,
     ChangePasswordComponent,
     BoardUsersComponent,
-    NavbarComponent,
   ],
   imports: [
     CommonModule,
@@ -30,13 +27,11 @@ import { AuthRoutingModule } from './auth/auth-routing.module';
     HttpClientModule,
     SharedModule,
     AuthRoutingModule,
-    //ErrorsRoutingModule,
+    ErrorsRoutingModule,
   ],
   exports: [
-    NavbarComponent,
   ],
   providers: [
-    authInterceptorProviders,
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
 })
