+++ FrontEnd/src/app/restaurant-management/restaurant-management-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { RolesEnum } from '../core/auth/enums/roles.enum';
-import { AuthGuard } from '../core/auth/guards/auth-guard';
+import { AuthGuard } from '../shared/guards/auth-guard';
 import { BoardRestaurantComponent } from './components/board-restaurant-management/board-restaurant-management.component';
 
 const routes: Routes = [
