+++ FrontEnd/src/app/shared/shared.module.ts
@@ -1,9 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { RouterModule } from '@angular/router';
 import { ButtonComponent } from './components/button/button.component';
 import { InputComponent } from './components/input/input.component';
+import { NavbarComponent } from './components/navbar/navbar.component';
 import { SimpleTableComponent } from './components/simple-table/simple-table.component';
+import { authInterceptorProviders } from './interceptors/auth.interceptor';
 
 
 @NgModule({
@@ -11,8 +14,10 @@ import { SimpleTableComponent } from './components/simple-table/simple-table.com
     SimpleTableComponent,
     InputComponent,
     ButtonComponent,
+    NavbarComponent,
   ],
   imports: [
+    RouterModule,
     CommonModule,
     FormsModule,
     ReactiveFormsModule,
@@ -21,6 +26,10 @@ import { SimpleTableComponent } from './components/simple-table/simple-table.com
     SimpleTableComponent,
     InputComponent,
     ButtonComponent,
+    NavbarComponent,
+  ],
+  providers: [
+    authInterceptorProviders,
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
 })
